pyproject.toml
[tool.poetry]
name = "pyvsystems_play"
version = "0.1.0"
description = ""
authors = ["bfmcneill <profitablepython@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.2"
numpy = "^1.19.2"
xlwings = "^0.20.7"
python-dotenv = "^0.14.0"
loguru = "^0.5.3"
jupyter = "^1.0.0"


[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
black = "^20.8b1"
pytest = "^6.1.0"
pytest-cov = "^2.10.1"
pytest-bpdb = "^0.1.4"
pytest-sugar = "^0.9.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
